+++ packages/backend/src/config.ts
@@ -25,6 +25,16 @@ export type Source = {
 		disableCache?: boolean;
 		extra?: { [x: string]: string };
 	};
+	dbReplications?: boolean;
+	dbSlaves?: {
+		host: string;
+		port: number;
+		db: string;
+		user: string;
+		pass: string;
+		disableCache?: boolean;
+		extra?: { [x: string]: string };
+	}[];
 	redis: {
 		host: string;
 		port: number;